Update package.json
allows use of the git branch via a package requirement. At the moment there is no release with the cancel order change that was just merged.
this allows a dev to spec.
 "orionx-sdk": "https://github.com/orionx-dev/orionx-sdk-js.git", 

in their package.json and use the latest master branch, as a work around till newer releases are rerolled.